refactor: migrate src js tree to extension domains
