+++ .prettierignore
@@ -2,4 +2,4 @@ node_modules
 build
 package-lock.json
 .idea
-src/js/fn/constants
+src/extension/core/constants
