+++ src/extension/core/collapse.ts
