+++ src/extension/core/copy.ts
