+++ src/extension/core/globals.ts
