+++ src/extension/services/CityMapRequestHandler.ts
