+++ src/extension/services/ClanBattleRequestHandler.ts
