+++ src/extension/services/ClanServiceRequestHandler.ts
