+++ src/extension/services/ConversationService.ts
@@ -15,10 +15,10 @@
 import { Tooltip, Alert, Popover } from 'bootstrap';
 import dayjs from 'dayjs';
 import { targets, targetsTopic } from '../index';
-import * as collapse from '../fn/collapse';
-import * as helper from '../fn/helper';
-import * as post_webstore from '../fn/post';
-import * as element from '../fn/AddElement';
+import * as collapse from '../core/collapse';
+import * as helper from '../core/helper';
+import * as post_webstore from '../core/post';
+import * as element from '../core/AddElement';
 import { setCurrentPercent } from './GreatBuildingsService';
 
 // targetsTopic = '🎯🎯 Battleground TARGETS 🎯🎯';
