+++ src/extension/services/GreatBuildingsRequestHandler.ts
