+++ src/extension/services/GuildBattlegroundRequestHandler.ts
