+++ src/extension/services/GuildBattlegroundSignalsRequestHandler.ts
