+++ src/extension/services/MiscRequestHandler.ts
