+++ src/extension/services/StartupRequestHandler.ts
