+++ src/extension/state/showOptions.ts
